+++ color/build.gradle.kts
@@ -23,7 +23,6 @@ kotlin {
 
         val commonMain by getting {
             dependencies {
-                api(project(":color"))
                 implementation(kotlin("stdlib"))
             }
         }
@@ -31,23 +30,14 @@ kotlin {
         val commonTest by getting {
             dependencies {
                 implementation(tuulbox.test())
-                implementation(kotlin("reflect"))
                 implementation(kotlin("test-common"))
                 implementation(kotlin("test-annotations-common"))
-            }
-        }
-
-        val androidMain by getting {
-            dependencies {
-                implementation("androidx.appcompat:appcompat:1.2.0")
             }
         }
 
         val androidTest by getting {
             dependencies {
                 implementation(kotlin("test-junit"))
-                implementation("androidx.test:core:1.3.0")
-                implementation("org.robolectric:robolectric:4.5.1")
             }
         }
