+++ color/gradle.properties
@@ -1,0 +1,1 @@
+POM_ARTIFACT_ID=color
