+++ color/src/androidMain/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.juul.krayon.color" />
