+++ sample/src/main/java/com/juul/krayon/sample/DemoActivity.kt
@@ -7,7 +7,7 @@ import com.juul.krayon.chart.data.ClusteredDataSet
 import com.juul.krayon.chart.data.toCluster
 import com.juul.krayon.chart.data.toRectangularDataSet
 import com.juul.krayon.chart.render.BarChartRenderer
-import com.juul.krayon.kanvas.nextColor
+import com.juul.krayon.color.nextColor
 import com.juul.krayon.sample.databinding.ActivityDemoBinding
 import kotlin.random.Random
